Big Update to Models
The database tables were intially inside with the models. However, when I created multiple many-to-many relationships, this became difficult to manage and produced many warnings. Since I could not verify whether the warnings were going to cause problem with the software, I played it safe and made each many-to-many relationship have a back_populates to explicitly assign a location to store the table in the different Models. This way we can have the three many-to-many relationships without fear of a collision. It was tested thouroughly but should try and test with more users to try and find inconsistencies.

The models were divided based on user and positions. This just cleans up the Model section and makes life easier when debugging/understanding the code.

app.py was adjusted to account for the relocation of the Models. I also attempted a module called Migrate but decided it was not worth integrating for I found another solution to placing Models in multiple files.