Added a new Model to store temporary deleted positions.
The DeletedPosition Model is connected to Student Model. When a Faculty deletes a position, we remove the application as
out design choice requires an existing Position for the Application to exist. Because we did not want to introduce a memory leak
due to deleting the Position but not the Application, we added a Model to temporarily flash to the Student when a position was deleted.
The DeletedPosition for the Student will be cleared after being notified [not implemented yet.]